Fix compile error with -Wshorten-64-to-32
The original patch was contributed by Alexis La Goutte

+++ examples/client.c
@@ -148,11 +148,11 @@ static ssize_t send_callback(nghttp2_session *session,
                              void *user_data)
 {
   struct Connection *connection;
-  ssize_t rv;
+  int rv;
   connection = (struct Connection*)user_data;
   connection->want_io = IO_NONE;
   ERR_clear_error();
-  rv = SSL_write(connection->ssl, data, length);
+  rv = SSL_write(connection->ssl, data, (int)length);
   if(rv < 0) {
     int err = SSL_get_error(connection->ssl, rv);
     if(err == SSL_ERROR_WANT_WRITE || err == SSL_ERROR_WANT_READ) {
@@ -177,11 +177,11 @@ static ssize_t recv_callback(nghttp2_session *session,
                              void *user_data)
 {
   struct Connection *connection;
-  ssize_t rv;
+  int rv;
   connection = (struct Connection*)user_data;
   connection->want_io = IO_NONE;
   ERR_clear_error();
-  rv = SSL_read(connection->ssl, buf, length);
+  rv = SSL_read(connection->ssl, buf, (int)length);
   if(rv < 0) {
     int err = SSL_get_error(connection->ssl, rv);
     if(err == SSL_ERROR_WANT_WRITE || err == SSL_ERROR_WANT_READ) {
